+++ CHANGELOG.md
@@ -16,6 +16,9 @@ All changed fall under either one of these types: `Added`, `Changed`, `Deprecate
 
 - stranded bigwigs are no longer inverted (forward containing reverse reads and vice-versa).
 - fix in `rename_sample` preventing the inversion of R1 and R2 FASTQs.
+- fixes in deseq2/utils.R
+  - the samples.tsv will now work with only 2 columns
+  - the samples.tsv column names will be stripped of excess whitespace, similar to the config.
 
 ## [0.5.5] - 2021-09-01